Add ability to view formation registrees in admin
